fix1

+++ Content.Server/_FarHorizons/Power/Generation/FissionGenerator/ReactorPartSystem.cs
@@ -25,8 +25,8 @@ public sealed class ReactorPartSystem : SharedReactorPartSystem
     /// <returns></returns>
     public GasMixture? ProcessGas(ReactorPartComponent reactorPart, Entity<NuclearReactorComponent> reactorEnt, AtmosDeviceUpdateEvent args, GasMixture inGas)
     {
-        if (reactorPart.RodType != (byte)ReactorPartComponent.RodTypes.GasChannel)
-            return null;
+        if (reactorPart.RodType != ReactorPartComponent.RodTypes.GasChannel)
+    return null;
 
         GasMixture? ProcessedGas = null;
         if (reactorPart.AirContents != null)
